backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via
    #   watchfiles
    #   watchgod
asgiref==3.5.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.2
    # via aioredis
attrs==22.1.0
    # via
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
autobahn==22.7.1
    # via daphne
automat==20.2.0
    # via twisted
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.0.0
    # via -r requirements.in
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
channels==3.0.5
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.4.1
    # via -r requirements.in
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.5
    # via
    #   click
    #   uvicorn
constantly==15.1.0
    # via twisted
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==37.0.4
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   social-auth-core
daphne==3.0.2
    # via channels
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.1.2
    # via
    #   -r requirements.in
    #   channels
    #   django-anymail
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   djangorestframework
    #   drf-yasg
django-anymail==8.6
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-debug-toolbar==3.6.0
    # via -r requirements.in
django-extensions==3.2.0
    # via -r requirements.in
django-redis-cache==3.0.1
    # via -r requirements.in
django-webpack-loader==1.6.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg[validation]==1.21.3
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
freezegun==1.2.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
hiredis==2.0.0
    # via aioredis
httptools==0.4.0
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
importlib-metadata==4.12.0
    # via markdown
incremental==21.3.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonschema==4.11.0
    # via swagger-spec-validator
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via
    #   -r requirements.in
    #   markdown-urlize
markdown-urlize==0.2.0
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.4
    # via channels-redis
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via drf-yasg
pillow==9.2.0
    # via -r requirements.in
prompt-toolkit==3.0.30
    # via click-repl
psycopg2==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pydot==1.4.2
    # via -r requirements.in
pyjwt==2.4.0
    # via social-auth-core
pyopenssl==22.0.0
    # via
    #   -r requirements.in
    #   twisted
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   packaging
    #   pydot
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   celery-redbeat
    #   freezegun
python-decouple==3.6
    # via -r requirements.in
python-dotenv==0.20.0
    # via uvicorn
python-jose==3.3.0
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.2.1
    # via
    #   celery
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via
    #   swagger-spec-validator
    #   uvicorn
redis==3.5.3
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   django-redis-cache
requests==2.28.1
    # via
    #   -r requirements.in
    #   coreapi
    #   django-anymail
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rsa==4.9
    # via python-jose
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sentry-sdk==1.9.5
    # via -r requirements.in
service-identity==21.1.0
    # via twisted
shortuuid==1.0.9
    # via -r requirements.in
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   ecdsa
    #   python-dateutil
    #   service-identity
    #   swagger-spec-validator
sniffio==1.2.0
    # via anyio
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
swagger-spec-validator==2.7.4
    # via drf-yasg
tenacity==8.0.1
    # via celery-redbeat
twisted[tls]==22.8.0
    # via daphne
txaio==22.2.1
    # via autobahn
typing-extensions==4.4.0
    # via twisted
tzdata==2022.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.11
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.18.2
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.16.1
    # via uvicorn
watchgod==0.8.2
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via uvicorn
werkzeug==2.2.2
    # via -r requirements.in
whitenoise==6.2.0
    # via -r requirements.in
zipp==3.8.1
    # via importlib-metadata
zope-interface==5.5.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
